+++ engine.py
@@ -1,7 +1,26 @@
 import torch
 from rich.progress import track
+from typing import Any, List, Tuple
 
-def train_fn(model, data_loader, optimizer, device):
+import torch
+import torch.nn as nn
+from torch.optim import Optimizer
+from torch.utils.data import DataLoader
+
+
+def train_fn(model: nn.Module, data_loader: DataLoader, optimizer: Optimizer, device: torch.device) -> float:
+    """
+    Trains the model on the provided data using the given optimizer.
+    
+    Parameters:
+        model (nn.Module): A PyTorch model to be trained.
+        data_loader (DataLoader): A PyTorch DataLoader object that provides batches of training data.
+        optimizer (Optimizer): A PyTorch optimizer used to update the model's parameters.
+        device (torch.device): A PyTorch device (such as 'cpu' or 'cuda') where the data and model should be loaded.
+    
+    Returns:
+        fin_loss (float): The average loss across all batches of training data.
+    """
     model.train()
     fin_loss = 0
 
@@ -16,7 +35,20 @@ def train_fn(model, data_loader, optimizer, device):
         fin_loss += loss.item()
     return fin_loss / len(data_loader)
 
-def eval_fn(model, data_loader, device):
+
+def eval_fn(model: nn.Module, data_loader: DataLoader, device: torch.device) -> Tuple[List[Any], float]:
+    """
+    Evaluates the model on the provided data.
+    
+    Parameters:
+        model (nn.Module): A PyTorch model to be evaluated.
+        data_loader (DataLoader): A PyTorch DataLoader object that provides batches of evaluation data.
+        device (torch.device): A PyTorch device (such as 'cpu' or 'cuda') where the data and model should be loaded.
+    
+    Returns:
+        fin_preds (list): A list of predictions made by the model on the evaluation data.
+        fin_loss (float): The average loss across all batches of evaluation data.
+    """
     model.eval()
     with torch.no_grad():
         fin_loss = 0